refactor: mvc 패턴에 맞게 패키지 변경
