+++ src/main/java/racingGame/controller/RacingGameController.java
@@ -1,14 +1,13 @@
-package racingGame;
+package racingGame.controller;
 
-import racingGame.game.GameResult;
-import racingGame.game.RacingGame;
-import racingGame.game.RandomStrategy;
-import racingGame.participant.Participants;
+import racingGame.domain.game.GameResult;
+import racingGame.domain.game.RacingGame;
+import racingGame.domain.game.RandomStrategy;
 import racingGame.view.InputView;
 import racingGame.view.RacingGameRequestDto;
 import racingGame.view.ResultView;
 
-public class RacingGameApplication {
+public class RacingGameController {
 
     public static void main(String[] args) {
         InputView inputView = new InputView();
