+++ src/main/java/racingGame/domain/car/Car.java
@@ -1,4 +1,4 @@
-package racingGame.car;
+package racingGame.domain.car;
 
 import java.util.Objects;
