+++ src/main/java/racingGame/domain/game/MovingRule.java
@@ -1,4 +1,4 @@
-package racingGame.game;
+package racingGame.domain.game;
 
 public interface MovingRule {
     public boolean isMove();
