+++ src/main/java/racingGame/domain/game/RandomStrategy.java
@@ -1,4 +1,4 @@
-package racingGame.game;
+package racingGame.domain.game;
 
 import java.util.Random;
