+++ src/main/java/racingGame/domain/participant/RoundScore.java
@@ -1,4 +1,4 @@
-package racingGame.participant;
+package racingGame.domain.participant;
 
 import java.util.List;
