+++ src/test/java/racingGame/domain/participant/ParticipantsTest.java
@@ -1,10 +1,10 @@
-package racingGame.participant;
+package racingGame.domain.participant;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingGame.car.Car;
-import racingGame.game.MovingRule;
+import racingGame.domain.car.Car;
+import racingGame.domain.game.MovingRule;
 
 import java.util.ArrayList;
 import java.util.Arrays;
